Ajuste parcial no update projetos e no state

+++ projecttreeMDI-main/src/main/java/br/projeto/command/SalvarProjetoDeEstimativaCommand.java
@@ -104,11 +104,14 @@ public class SalvarProjetoDeEstimativaCommand implements Command{
             projetoDeEstimativaModel.setId(projetoId);
         
             projetoDeEstimativaPresenter.getProjetoDeEstimativaRepository().update(projetoDeEstimativaModel);
-        
-            projetoDeEstimativaPresenter.getProjetoFuncionalidadesPersonalizadasRepository().deleteByProjetoDeEstimativa(projetoDeEstimativaModel);
-        
-            projetoDeEstimativaPresenter.getPerfilProjetoIntermediariaRepository().deleteByProjeto(projetoId);
-        
+            
+            if(!projetoDeEstimativaPresenter.getPerfilProjetoIntermediariaRepository().findByProjeto(projetoId).isEmpty()){
+                projetoDeEstimativaPresenter.getPerfilProjetoIntermediariaRepository().deleteByProjeto(projetoId);
+            }
+            
+            if(!projetoDeEstimativaPresenter.getProjetoFuncionalidadesPersonalizadasRepository().findByProjetoEstimativa(projetoDeEstimativaModel).isEmpty()){
+                projetoDeEstimativaPresenter.getProjetoFuncionalidadesPersonalizadasRepository().deleteByProjetoDeEstimativa(projetoDeEstimativaModel);
+            }
         }else{
             projetoDeEstimativaPresenter.getProjetoDeEstimativaRepository().insert(projetoDeEstimativaModel);
         }
